+++ semcore/drag-and-drop/src/DragAndDrop.tsx
@@ -7,35 +7,9 @@ import useEnhancedEffect from '@semcore/core/lib/utils/use/useEnhancedEffect';
 import { Box, ScreenReaderOnly } from '@semcore/flex-box';
 import React from 'react';
 
-import type { DropZoneProps } from './index';
+import type { DragAndDropProps, DropZoneProps } from './index';
 import style from './style/drag-and-drop.shadow.css';
 import { localizedMessages } from './translations/__intergalactic-dynamic-locales';
-
-type AsProps = {
-  /**
-   * @deprecated don't use this prop
-   */
-  theme?: 'dark' | 'default';
-  /**
-   * @deprecated use `onDnD` instead
-   */
-  onSwapDraggable?: (draggableNode: React.ReactNode, droppableNode: React.ReactNode) => void;
-  /**
-   * @deprecated use `onDnD` instead
-   */
-  onInsertDroppable?: (draggableNode: React.ReactNode, droppableNode: React.ReactNode) => void;
-  /**
-   * Contolled drag and drop handler
-   */
-  onDnD: (dndData: { fromIndex: number; fromId: string; toIndex: number; toId: string }) => void;
-  /**
-   * Index of id that indicates item that is currently under the user focus in case of real browser focus cannot be used.
-   * When provided, drag and drop listens to whole page keyboard events. Doesn't provide `onCustomFocusChange` callback.
-   */
-  customFocus?: number | string;
-  getI18nText: (messageId: string, values?: { [key: string]: string | number }) => string;
-  uid: string;
-};
 
 type AttachDetails = {
   index: number;
@@ -63,6 +37,7 @@ type State = {
   dragging: null | {
     index: number;
     initialItemsRects: Array<{ x: number; y: number; width: number; height: number } | undefined>;
+    placeholder: HTMLElement | null;
   };
   dragOver: number | null;
   hideHoverEffect: boolean;
@@ -74,9 +49,9 @@ type State = {
 
 type A11yHintKeys = keyof typeof localizedMessages.en;
 
-class DragAndDropRoot extends Component<AsProps, {}, State> {
+class DragAndDropRoot extends Component<DragAndDropProps, {}, State, typeof DragAndDropRoot.enhance> {
   static displayName = 'DragAndDrop';
-  static enhance = [i18nEnhance(localizedMessages), uniqueIDEnhance()];
+  static enhance = [i18nEnhance(localizedMessages), uniqueIDEnhance()] as const;
   static defaultProps = {
     theme: 'default',
     i18n: localizedMessages,
@@ -99,10 +74,10 @@ class DragAndDropRoot extends Component<AsProps, {}, State> {
   containerRef = React.createRef<HTMLDivElement>();
   clearA11yHintTimeout = 0;
 
-  handleItemDragStart = (index: number) => {
+  handleItemDragStart = (index: number, event: Event) => {
     const { items } = this.state;
     const currentItem = items[index];
-    if (!currentItem) return;
+    if (!currentItem || !(event.target instanceof HTMLElement)) return;
 
     const itemText =
       currentItem.node.getAttribute('aria-label') || currentItem.node.textContent || `${index + 1}`;
@@ -121,16 +96,60 @@ class DragAndDropRoot extends Component<AsProps, {}, State> {
       zoneName: zoneName || '',
     });
 
+    let placeholder: HTMLDivElement | null = null;
+
+    if (event.type === 'dragstart') {
+      placeholder = document.createElement('div');
+      placeholder.style.backgroundColor = 'var(--intergalactic-bg-primary-neutral-hover, #f4f5f9)';
+      placeholder.style.width = event.target.offsetWidth + 'px';
+      placeholder.style.height = event.target.offsetHeight + 'px';
+
+      const target = event.target;
+
+      event.target.parentNode?.insertBefore(placeholder, target.nextSibling);
+      target.style.left = `${target.offsetLeft}px`;
+      target.style.top = `${target.offsetTop}px`;
+      target.style.position = 'absolute';
+
+      setTimeout(() => { // because FF and safari can't create visible draggableImage without timeout
+        target.style.opacity = '0';
+      });
+    }
+
+    const yOffset = this.asProps.scrollableContainerRef?.current?.scrollTop ?? 0;
+    const xOffset = this.asProps.scrollableContainerRef?.current?.scrollLeft ?? 0;
+
     this.setState((prevState: State) => ({
       dragging: {
         index,
-        initialItemsRects: prevState.items.map((item) => item?.node.getBoundingClientRect()),
+        initialItemsRects: prevState.items.map((item) => {
+          if (!item) return;
+
+          const rect = item.node.getBoundingClientRect();
+
+          return {
+            width: rect.width,
+            height: rect.height,
+            x: rect.x + xOffset,
+            y: rect.y + yOffset,
+          };
+        }),
+        placeholder,
       },
     }));
   };
 
   handleItemDragEnd = (event: DragEvent) => {
     event.preventDefault();
+
+    const { items, dragging } = this.state;
+    const draggingItem = dragging ? items[dragging.index] : undefined;
+    const placeholder = dragging?.placeholder;
+
+    if (!draggingItem || !placeholder) return;
+
+    this.clearPlaceholder(draggingItem.node, placeholder);
+
     this.setState({
       dragging: null,
       dragOver: null,
@@ -143,19 +162,23 @@ class DragAndDropRoot extends Component<AsProps, {}, State> {
     event.preventDefault();
     const { items, dragging, dragOver } = this.state;
     if (!dragging) return;
+
+    const yOffset = this.asProps.scrollableContainerRef?.current?.scrollTop ?? 0;
+    const xOffset = this.asProps.scrollableContainerRef?.current?.scrollLeft ?? 0;
+
     const itemIndex = dragging.initialItemsRects.findIndex(
       (rect) =>
         rect &&
-        event.clientX > rect.x &&
-        event.clientX < rect.x + rect.width &&
-        event.clientY > rect.y &&
-        event.clientY < rect.y + rect.height,
+        event.clientX + xOffset > rect.x &&
+        event.clientX + xOffset < rect.x + rect.width &&
+        event.clientY + yOffset > rect.y &&
+        event.clientY + yOffset < rect.y + rect.height,
     );
     const currentItem = items[itemIndex];
-    const draggingItem = items[dragging.index];
+    const draggingItem = dragging.placeholder;
     if (!currentItem || !draggingItem || itemIndex === dragOver) return;
 
-    const node = currentItem.isDropZone ? draggingItem.node : currentItem.node;
+    const node = currentItem.isDropZone ? draggingItem : currentItem.node;
 
     const itemText = node.getAttribute('aria-label') || node.textContent || `${itemIndex + 1}`;
     const zoneName = currentItem.zoneName;
@@ -181,7 +204,7 @@ class DragAndDropRoot extends Component<AsProps, {}, State> {
     this.setState({ dragOver: itemIndex }, this.swapElements);
   };
 
-  handleItemDrop = () => {
+  handleItemDrop = (event: Event) => {
     const { onDnD } = this.asProps;
     if (!onDnD) return;
 
@@ -223,6 +246,12 @@ class DragAndDropRoot extends Component<AsProps, {}, State> {
       reversedScaling: false,
     });
     if (dragging && items[dragging.index]) {
+      const placeholder = dragging.placeholder;
+
+      if (placeholder) {
+        this.clearPlaceholder(draggingItem.node, placeholder);
+      }
+
       const fromNode = items[dragging.index];
       if (fromNode) {
         onDnD({
@@ -288,13 +317,23 @@ class DragAndDropRoot extends Component<AsProps, {}, State> {
     }
   };
 
+  clearPlaceholder(node: HTMLElement, placeholder: HTMLElement) {
+    this.containerRef.current?.insertBefore(node, placeholder);
+    node.style.left = '0';
+    node.style.top = '0';
+    node.style.opacity = '1.0';
+    node.style.position = 'relative';
+
+    placeholder.remove();
+  }
+
   swapElements = () => {
     const { items, dragging, dragOver } = this.state;
     const draggingIndex = dragging?.index ?? null;
 
     if (draggingIndex === null || dragOver === null) return;
 
-    const node = items[draggingIndex]?.node;
+    const node = dragging?.placeholder ?? items[draggingIndex]?.node;
     const dragNode = items[dragOver]?.node;
 
     if (!node || !dragNode) return;
@@ -309,6 +348,11 @@ class DragAndDropRoot extends Component<AsProps, {}, State> {
     }
 
     node.focus();
+
+    if (dragging?.placeholder && items[draggingIndex]?.node) {
+      items[draggingIndex].node.style.left = `${node.offsetLeft}px`;
+      items[draggingIndex].node.style.top = `${node.offsetTop}px`;
+    }
   };
 
   handleItemKeyDown = (event: KeyboardEvent, index: number) => {
@@ -325,7 +369,7 @@ class DragAndDropRoot extends Component<AsProps, {}, State> {
       event.preventDefault();
       event.stopPropagation();
       if (this.state.dragging) {
-        this.handleItemDrop();
+        this.handleItemDrop(event);
         this.setState({
           dragging: null,
           dragOver: null,
@@ -334,7 +378,7 @@ class DragAndDropRoot extends Component<AsProps, {}, State> {
           hideHoverEffect: true,
         });
       } else if (this.state.items[index]?.draggingAllowed) {
-        this.handleItemDragStart(index);
+        this.handleItemDragStart(index, event);
         this.setState({ keyboardDraggingIndex: index, animatedScaling: index });
         this.updateItemScaling();
       }
@@ -421,7 +465,7 @@ class DragAndDropRoot extends Component<AsProps, {}, State> {
     this.updateItemScaling();
   };
 
-  makeItemDragStartHandler = (index: number) => () => this.handleItemDragStart(index);
+  makeItemDragStartHandler = (index: number) => (e: DragEvent) => this.handleItemDragStart(index, e);
   makeItemKeyDownHandler = (index: number) => (event: KeyboardEvent) =>
     this.handleItemKeyDown(event, index);
 
@@ -444,6 +488,8 @@ class DragAndDropRoot extends Component<AsProps, {}, State> {
       dark: this.asProps.theme === 'dark',
       hideHoverEffect: this.state.hideHoverEffect,
       animatedScaling: index === this.state.animatedScaling,
+      active: index === this.state.dragging?.index ? 'true' : 'false',
+      isDragging: this.state.dragging !== null,
     };
   }
 
@@ -518,7 +564,7 @@ class DragAndDropRoot extends Component<AsProps, {}, State> {
     document.removeEventListener('keydown', this.handlePageKeyDown, { capture: true });
   }
 
-  componentDidUpdate(prevProps: AsProps) {
+  componentDidUpdate(prevProps: DragAndDropProps) {
     if (prevProps.customFocus !== this.asProps.customFocus) {
       const itemIndex = this.getCustomFocusItemIndex(this.asProps.customFocus);
       if (this.state.items[itemIndex!]) this.handleItemFocus();
@@ -531,6 +577,7 @@ class DragAndDropRoot extends Component<AsProps, {}, State> {
     const { a11yHint } = this.state;
     const context = { attach, detach };
     const { getI18nText, uid } = this.asProps;
+    const SDnDContainer = Root;
 
     return sstyled(this.asProps.styles)(
       <DragAndDropContext.Provider value={context}>
@@ -539,7 +586,7 @@ class DragAndDropRoot extends Component<AsProps, {}, State> {
             {a11yHint}
           </SA11yHint>
         )}
-        <Root render={Box} role='group' ref={this.containerRef} />
+        <SDnDContainer render={Box} role='group' ref={this.containerRef} />
         <ScreenReaderOnly id={`describe-draggable-${uid}`} aria-hidden='true'>
           {getI18nText('describe', { control: 'Space' })}
         </ScreenReaderOnly>