+++ semcore/time-picker/__tests__/time-picker.axe-test.ts
@@ -2,8 +2,21 @@ import { e2eStandToHtml } from '@semcore/testing-utils/e2e-stand';
 import { expect, test, getAccessibilityViolations } from '@semcore/testing-utils/playwright';
 
 test.describe('Time picker', () => {
-  test('Basic', async ({ page }) => {
-    const standPath = 'stories/components/time-picker/docs/examples/expanded_access_to_all_the_components.tsx';
+  test('Datepicker and timepicker', async ({ page }) => {
+    const standPath = 'stories/patterns/ux-patterns/form/docs/examples/datepicker-and-timepicker.tsx';
+    const htmlContent = await e2eStandToHtml(standPath, 'en');
+
+    await page.setContent(htmlContent);
+
+    const violations = await getAccessibilityViolations({ page });
+
+    expect(violations).toEqual([]);
+  });
+});
+
+test.describe('UX pattern', () => {
+  test('Datepicker and timepicker', async ({ page }) => {
+    const standPath = 'stories/patterns/ux-patterns/form/docs/examples/datepicker-and-timepicker.tsx';
     const htmlContent = await e2eStandToHtml(standPath, 'en');
 
     await page.setContent(htmlContent);