+++ stories/components/data-table/docs/data-table.docs.stories.tsx
@@ -19,6 +19,8 @@ import EmptyTableExample from './examples/empty-table';
 import ExportInImageExample from './examples/export-in-image';
 import FixedColumnsExample from './examples/fixed-columns';
 import FixedHeaderExample from './examples/fixed-header';
+import type { LimitedModeExampleProps } from './examples/limited-mode';
+import LimitedModeExample, { limitedModeDefaultProps } from './examples/limited-mode';
 import MultiLevelHeaderExample from './examples/multi-level-header';
 import PaginationExample from './examples/pagination';
 import RowThemesExample from './examples/row-themes';
@@ -192,4 +194,9 @@ export const VirtualScrollInTable: Story = {
 
 export const VirtualScrollInTableDifferentHeight: Story = {
   render: VirtualScrollInTableDifferentHeightExample,
+};
+
+export const LimitedMode: StoryObj<LimitedModeExampleProps> = {
+  render: LimitedModeExample,
+  args: limitedModeDefaultProps,
 };