+++ stories/components/data-table/tests/examples/scroll-tests/fixed-column-with-d-ff-width.tsx
@@ -17,7 +17,6 @@ const Demo = () => {
           { name: 'cpc', children: 'CPC', gtcWidth: '300px' },
           { name: 'cpc', children: 'CPC', gtcWidth: '100px' },
           { name: 'vol', children: 'Vol.', gtcWidth: '80px', fixed: 'right' },
-
         ]}
       />