+++ stories/patterns/ux-patterns/form/docs/Form.stories.tsx
@@ -1,18 +1,18 @@
 import type { Meta, StoryObj } from '@storybook/react-vite';
 
 import { DefaultLogInFormTest } from './__tests__/DefaultLogInForm.test';
-import CheckboxAndRadioExample from './examples/checkbox-and-radio';
 import DatepickerAndTimepickerExample from './examples/datepicker-and-timepicker';
 import DefaultLogInFormExample from './examples/default-log-in-form';
 import InputTagsAndSelectExample from './examples/inputtags-and-select';
+import RadioAndSelectExample from './examples/radio-and-select';
 import { playWrapper } from '../../../../utils/playWrapper';
 
 const meta: Meta = {
   title: 'Patterns/UX Patterns/Form',
 };
 
-export const CheckboxAndRadio: StoryObj = {
-  render: CheckboxAndRadioExample,
+export const RadioAndSelect: StoryObj = {
+  render: RadioAndSelectExample,
 };
 
 export const DatepickerAndTimepicker: StoryObj = {