+++ semcore/d3-chart/CHANGELOG.md
@@ -2,7 +2,17 @@
 
 CHANGELOG.md standards are inspired by [keepachangelog.com](https://keepachangelog.com/en/1.0.0/).
 
-## [16.1.5] - 2025-10-03
+## [16.2.1] - 2025-10-17
+
+### Fixed
+
+- Chart content remains visible after unchecking single legend item in Bar, Horizontal Bar, Histogram, and Stacked Horizontal Bar charts.
+
+## [16.2.0] - 2025-10-03
+
+### Changed
+
+- Focus styles for `Plot`.
 
 ### Fixed