+++ semcore/dropdown/CHANGELOG.md
@@ -2,6 +2,12 @@
 
 CHANGELOG.md standards are inspired by [keepachangelog.com](https://keepachangelog.com/en/1.0.0/).
 
+## [16.1.0] - 2025-10-03
+
+### Changed
+
+- Styles for `:focus`.
+
 ## [16.0.9] - 2025-09-17
 
 ### Changed