+++ semcore/feature-highlight/CHANGELOG.md
@@ -2,6 +2,18 @@
 
 CHANGELOG.md standards are inspired by [keepachangelog.com](https://keepachangelog.com/en/1.0.0/).
 
+## [16.2.1] - 2025-10-17
+
+### Added
+
+- Types for `NoticeFH`.
+
+## [16.2.0] - 2025-10-03
+
+### Changed
+
+- Styles for `:focus-visible`.
+
 ## [16.1.1] - 2025-09-17
 
 ### Changed